Protect users passwords on failed attempt
Change is to stop password from being logged in database in clear text. If a user enters wrong password currently that password is stored in clear text in the database in TABLE: members_users_login. This is bad practice since, say the user forgets his password and tries a few of his regular passwords for other websites. The user has unknowingly stored a legitimate password and email address in a database and doesn't even know what has gone on.