Remove "Scammers Warming" footer message
Looks a bit old-fashioned and ugly.

Every site owner is more than welcome to add it on their website, but I don't think it's a good idea to add this warning by default on every website.
Especially if the website is just a simple community or a social network, and not a dating website.